various fixes

+++ src/einundzwanzig.py
@@ -84,13 +84,13 @@ def getInvoice(amt: int, memo: str) -> str:
     dict = json.loads(response)
     return dict["lightning_pay_request"]
 
-def createQR(text: str, chatid: str):
+def createQR(text: str, chatid: int):
     """
     Creates a QR Code with the given text and saves it as a png file in the current directory
     """
 
     img = qrcode.make(text)
-    img.save(chatid + '.png')
+    img.save(f"{chatid}.png")
 
 
 def shoutout(update: Update, context: CallbackContext) -> int:
@@ -101,13 +101,14 @@ def shoutout(update: Update, context: CallbackContext) -> int:
     chat = update.effective_chat
     if chat != None and chat.type == Chat.PRIVATE:
         update.message.reply_text(dedent('''
-        Hi! Du möchtest also einen Shout-Out bei Einundzwanzig kaufen? Toll! Du kannst den Vorgang jederzeit mit /cancel abbrechen\n\n
-        Bitte nenne jetzt die Menge an Satoshis die du spenden möchtest (Bspw: "21000"). Beachte, dass lediglich Spenden
-        größer als 21.000 Satoshis in der nächsten Newsepisode vorgelesen werden 
-        '''))
+        Hi! Du möchtest also einen Shoutout bei Einundzwanzig kaufen? Toll!\n
+        Bitte nenne zuerst die Menge an Satoshis die du spenden möchtest (Bspw: 21000). Beachte, dass lediglich Spenden
+        größer als <b>21.000</b> Satoshis in der nächsten Newsepisode vorgelesen werden.\n
+        Du kannst den Vorgang jederzeit mit /cancel abbrechen.
+        '''), parse_mode='HTML')
         return SHOUTOUT_AMOUNT
     else:
-        update.message.reply_text('Shoutouts können nur im privaten Chat mit dem Bot gesendet werden. Bitte beginne einen direkten Chat mit @einundzwanzigbot !')
+        update.message.reply_text('Shoutouts können nur im privaten Chat mit dem Bot gesendet werden. Bitte beginne einen direkten Chat mit @einundzwanzigbot.')
         return ConversationHandler.END
 
 
@@ -115,13 +116,29 @@ def memo(update: Update, context: CallbackContext) -> int:
     """
     Stores the amount and asks for the donation-message (memo)
     """
-    
-    text = update.message.text
-    context.user_data['amount'] = int(text)
+
+    if context.user_data == None:
+        update.message.reply_text(text='Es ist ein Fehler aufgetreten, bitte versuche es später noch mal')
+        logging.error('context.user_data is None')
+        return ConversationHandler.END
+ 
+    try:
+        sat_amount = int(update.message.text)
+
+        if sat_amount <= 0:
+            update.message.reply_text(text='Der Betrag darf nicht kleiner oder gleich 0 sein!')
+            return SHOUTOUT_AMOUNT
+
+        context.user_data['amount'] = sat_amount
+    except:
+        update.message.reply_text(text='Bitte gib eine korrekte Anzahl von sats ein!')
+        return SHOUTOUT_AMOUNT
+
     update.message.reply_text(dedent(f'''
-    Alles klar! Du möchtest also {context.user_data["amount"]} Satoshis spenden.\n\n
-    Bitte füge deiner Spende noch eine Nachricht (max. 140 Zeichen) hinzu. Spenden über 21.000 Satoshi werden in der nächsten Newsepisode vorgelesen!
+    Alles klar! Du möchtest also {sat_amount} Satoshis spenden.\n
+    Bitte füge deiner Spende noch eine Nachricht (max. 140 Zeichen) hinzu.
     '''))
+
     return SHOUTOUT_MEMO
 
 
@@ -130,45 +147,63 @@ def invoice(update: Update, context: CallbackContext) -> int:
     Stores the message and returns the lightning invoice + qr code.
     """
     memo = update.message.text
-    context.user_data['memo'] = memo
-    amount = context.user_data['amount']
-    if len(memo) > 140:
-        update.message.reply_text(text='Der Shoutout darf nicht länger als 140 Zeichen sein. Bitte beginne von vorne')
+
+    # If we don't have user data we need to abort
+    if context.user_data == None:
+        update.message.reply_text(text='Es ist ein Fehler aufgetreten, bitte versuche es später noch mal')
+        logging.error('context.user_data is None')
         return ConversationHandler.END
+
+    if update.effective_chat == None:
+        update.message.reply_text(text='Es ist ein Fehler aufgetreten, bitte versuche es später noch mal')
+        logging.error('update.effecitve_chat is None')
+        return ConversationHandler.END
+
+    context.user_data['memo'] = memo
+
+    sat_amount: int = context.user_data['amount']
+
+    if len(memo) > 140:
+        update.message.reply_text(text=f'Der Shoutout darf nicht länger als 140 Zeichen sein. Aktuelle Länge: {len(memo)} Zeichen.')
+        return SHOUTOUT_MEMO
     else:
         
         try:
-            invoice = getInvoice(amount, memo)     
+            invoice = getInvoice(sat_amount, memo)     
         except Exception as e:
             context.bot.send_message(chat_id=update.effective_chat.id, text= f'Fehler beim Erstellen der Invoice! Bitte versuche es später nochmal!')
             logging.error(f'Error while trying to generate invoice: {e}')
-            return
+            return ConversationHandler.END
         
         try:
             createQR(invoice, update.effective_chat.id)
         except Exception as e:
             context.bot.send_message(chat_id=update.effective_chat.id, text= f'Fehler beim Erstellen des QR-Codes! Bitte versuche es später nochmal!')
             logging.error(f'Error while trying to generate QR code: {e}')
-            return
+            return ConversationHandler.END
 
         shoutoutMessage = dedent(f'''
         <b>Dein Shoutout</b>
-        Betrag: {amount} Satoshis
+        Betrag: {sat_amount} Satoshis
         Memo: {memo}
+
+        Sobald die Invoice bezahlt wurde ist der Vorgang abgeschlossen.
+        Deine Nachricht und deine sats sind dann bei uns angekommen.
         ''')
 
-        context.bot.send_message(chat_id=update.effective_chat.id, text=shoutoutMessage)
+        context.bot.send_message(chat_id=update.effective_chat.id, text=shoutoutMessage, parse_mode='HTML')
         context.bot.send_photo(chat_id=update.effective_chat.id, photo=open(f'{update.effective_chat.id}.png', 'rb'), caption=str(invoice).lower())
         
         try:
             os.remove(f'{update.effective_chat.id}.png')
         except:
             logging.error(f'ERROR: QR-code file {update.effective_chat.id}.png could not be deleted')
+
         return ConversationHandler.END
 
 
 def cancel(update: Update, context: CallbackContext) -> int:
     """Cancels and ends the conversation."""
-    update.message.reply_text('Shoutout Abgrebrochen! Bis zum nächsten mal!')
+    update.message.reply_text('Shoutout abgebrochen.')
     
     return ConversationHandler.END
